+++ src/scalecast/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.17.12'
+__version__ = '0.17.13'
 
 from .util import metrics
 import inspect